updating functions
added an API caller that extracts information from the LCP Dictionary, a function that extracts explanations for each experimental condition, and made the exp. conditions that are not Time or Code into Functional Components, each with properties describing their amounts and units. They show up only if they are present in the Sample, ie do not have a value of '0'. Also updated run_from_terminal to account for the new functions.